+++ packages/enhanced/test/config_test.dart
@@ -300,9 +300,8 @@ void main() {
       expect(
           explained,
           equals('[CssSizing:height≥0.0,height=auto,width≥0.0,width=auto,child='
-              '[Image:image=CachedNetworkImageProvider("http://base.com/path/image.png", scale: 1.0),'
-              'semanticLabel=image dot png'
-              ']]'));
+              '[CachedNetworkImage:imageUrl=http://base.com/path/image.png]'
+              ']'));
     });
   });
