+++ packages/fwfh_cached_network_image/lib/src/cached_network_image_factory.dart
@@ -17,14 +17,14 @@ mixin CachedNetworkImageFactory on WidgetFactory {
       fit: BoxFit.fill,
       imageUrl: url,
       progressIndicatorBuilder: (context, _, progress) => imageLoadingBuilder(
-          context,
-          const SizedBox.shrink(),
-          ImageChunkEvent(
-            cumulativeBytesLoaded: progress.downloaded,
-            expectedTotalBytes: progress.totalSize,
-          ),
-          src,
+        context,
+        const SizedBox.shrink(),
+        ImageChunkEvent(
+          cumulativeBytesLoaded: progress.downloaded,
+          expectedTotalBytes: progress.totalSize,
         ),
+        src,
+      ),
     );
   }
 }
