+++ packages/fwfh_cached_network_image/test/cached_network_image_factory_test.dart
@@ -1,25 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
-import 'package:fwfh_cached_network_image/fwfh_cached_network_image.dart';
 
-import '../../core/test/_.dart' as helper;
-
-Future<String> explain(
-  WidgetTester tester,
-  String html, {
-  bool webView = true,
-}) async =>
-    helper.explain(
-      tester,
-      null,
-      hw: HtmlWidget(
-        html,
-        key: helper.hwKey,
-        factoryBuilder: () => _WidgetFactory(),
-      ),
-    );
-
-class _WidgetFactory extends WidgetFactory with CachedNetworkImageFactory {}
+import '_.dart';
 
 void main() {
   testWidgets('renders IMG tag', (WidgetTester tester) async {
@@ -30,7 +11,7 @@ void main() {
     expect(
       explained,
       equals('[CssSizing:$sizingConstraints,child='
-          '[Image:image=CachedNetworkImageProvider("$src", scale: 1.0)]'
+          '[CachedNetworkImage:imageUrl=$src]'
           ']'),
     );
   });
